RedditStream/TrustedZone/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0
pandas==2.0.1
numpy==1.23.5
google-cloud-storage==2.5.0